working on mount

+++ kernel/arch/amd64/include/arch_proto.h
@@ -279,7 +279,7 @@
 	unsigned long do_execve(stack_frame_s * curr_context, char *exec_filename, char *argv[], char *envp[]);
 	unsigned long do_exit(unsigned long exit_code);
 	void schedule(void);
-	int kernel_thread(unsigned long (* fn)(unsigned long), unsigned long arg, unsigned long flags);
+	unsigned long kernel_thread(unsigned long (* fn)(unsigned long), unsigned long arg, unsigned long flags);
 	unsigned long init(unsigned long arg);
 	int user_thread_test(unsigned long (* fn)(unsigned long), unsigned long arg, unsigned long flags);
 	unsigned long user_func(unsigned long arg);
